Add property filtration by period
