+++ TravelWithPleasureHub/src/main/java/com/kh021j/travelwithpleasurehub/model/Property.java
@@ -1,4 +1,4 @@
-package com.kh021j.travelwithpleasurehub.models;
+package com.kh021j.travelwithpleasurehub.model;
 
 
 import lombok.AllArgsConstructor;
@@ -6,31 +6,35 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-import java.time.LocalDate;
 
 @Entity
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class Application {
+public class Property {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
+    private String title;
+
+    private String description;
+
     @ManyToOne
-    @JoinColumn(name = "property_id")
-    private Property property;
+    @JoinColumn(name = "property_type_id")
+    private PropertyType propertyType;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
     private User userTable;
 
-    private LocalDate rentSince;
+    private String locality;
 
-    private LocalDate rentUntil;
+    private String address;
 
-    private String applicationText;
+    private Integer price;
 
-    private Boolean is_approved;
+    private String pathToPhoto;
+
 }
