+++ TravelWithPleasureHub/src/main/java/com/kh021j/travelwithpleasurehub/model/PropertyReview.java
@@ -1,0 +1,32 @@
+package com.kh021j.travelwithpleasurehub.model;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Entity
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class PropertyReview {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+
+    private String reviewText;
+
+    private Integer rate;
+
+    @ManyToOne
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    @ManyToOne
+    @JoinColumn(name = "property_id")
+    private Property property;
+
+}
