--- TravelWithPleasureHub/src/main/java/com/kh021j/travelwithpleasurehub/models/Property.java
@@ -1,38 +1,0 @@
-package com.kh021j.travelwithpleasurehub.models;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Property {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    private String title;
-
-    private String description;
-
-    @ManyToOne
-    @JoinColumn(name = "property_type_id")
-    private PropertyType propertyType;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    private User userTable;
-
-    private String locality;
-
-    private String address;
-
-    private Integer price;
-
-}
