--- TravelWithPleasureHub/src/main/java/com/kh021j/travelwithpleasurehub/repositories/ApplicationRepository.java
@@ -1,11 +1,0 @@
-package com.kh021j.travelwithpleasurehub.repositories;
-
-import com.kh021j.travelwithpleasurehub.models.Application;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-
-@Repository
-public interface ApplicationRepository extends JpaRepository<Application, Integer> {
-
-}
