--- TravelWithPleasureHub/src/main/java/com/kh021j/travelwithpleasurehub/repositories/PropertyAvailabilityRepository.java
@@ -1,10 +1,0 @@
-package com.kh021j.travelwithpleasurehub.repositories;
-
-import com.kh021j.travelwithpleasurehub.models.PropertyAvailability;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PropertyAvailabilityRepository extends JpaRepository<PropertyAvailability, Integer> {
-
-}
