--- TravelWithPleasureHub/src/main/java/com/kh021j/travelwithpleasurehub/repositories/UserReviewRepository.java
@@ -1,10 +1,0 @@
-package com.kh021j.travelwithpleasurehub.repositories;
-
-import com.kh021j.travelwithpleasurehub.models.UserReview;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserReviewRepository extends JpaRepository<UserReview, Integer> {
-
-}
